Переписала реализацию интерфейса в модель Users, добавила настройки конфига
